restructure fft directory
